Fix clang-tidy narrowing-conversion error in setFramesCompleted call
framesCompleted is uint32_t but RenderRequest::setFramesCompleted takes
int; cast explicitly instead of relying on an implementation-defined
narrowing conversion.

+++ scheduler/src/scheduler.cpp
@@ -156,7 +156,7 @@ bool Scheduler::markRenderCompleted(const std::string& workerJobId) {
         }
     }
 
-    render->setFramesCompleted(framesCompleted);
+    render->setFramesCompleted(static_cast<int>(framesCompleted));
 
     if (allFramesDone) {
         render->setDownloadLinks(orderedLinks);
